this is a terribly large risky commit, but not really many good ways to chop it up.  this converts all calls to the now-deprecated BU_PTBL_END to the type-safe and null-safe BU_PTBL_LEN macro.  that consequently requires propagating a slew of size_t type changes.  consequently found and fixed a bunch of 1-byte memory-overruns in the iges converter while converting types too.  regression testing passes on mac.
svn:revision:68052
svn:branch:trunk
svn:account:brlcad

+++ src/conv/iges/get_cnurb_curve.c
@@ -25,14 +25,14 @@ struct edge_g_cnurb *
 Get_cnurb_curve(int curve_de, int *linear)
 {
     int i;
-    int curve;
+    size_t curve;
     struct edge_g_cnurb *crv;
 
     *linear = 0;
 
     curve = (curve_de - 1)/2;
     if (curve >= dirarraylen) {
-	bu_log("Get_cnurb_curve: DE=%d is too large, dirarraylen = %d\n", curve_de, dirarraylen);
+	bu_log("Get_cnurb_curve: DE=%zu is too large, dirarraylen = %zu\n", curve_de, dirarraylen);
 	return (struct edge_g_cnurb *)NULL;
     }
